fix: restore main CI after context provenance update
