+++ src/agents/embedded-agent-runner/run/attempt-normalization.ts
@@ -164,7 +164,9 @@ export async function normalizeEmbeddedRunAttempt(input: {
   const promptCacheLastCallUsage = normalizeUsage(attempt.promptCache?.lastCallUsage as UsageLike);
   const callUsage = resolveLatestCallUsage({
     currentAttemptCandidates: [currentAttemptAssistantUsage, promptCacheLastCallUsage],
-    carriedCandidates: [input.lastRunPromptUsage, lastAssistantUsage],
+    // The transcript assistant is newer than the prior retry snapshot. Its unavailable sentinel
+    // must win so an older context total cannot be revived after a provider omits fresh usage.
+    carriedCandidates: [lastAssistantUsage, input.lastRunPromptUsage],
   });
   const attemptUsage = attempt.attemptUsage ?? callUsage.currentAttempt;
   mergeUsageIntoAccumulator(input.usageAccumulator, attemptUsage);
