+++ src/agents/sessions/agent-session-loop-correctness.test.ts
@@ -6,233 +6,20 @@ import {
   type SimpleStreamOptions,
 } from "openclaw/plugin-sdk/llm";
 import { Type } from "typebox";
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-
-const streamMocks = vi.hoisted(() => ({
-  streamSimple: vi.fn(),
-}));
-
+import { describe, expect, it, vi } from "vitest";
 import type { AgentTool } from "../runtime/index.js";
-import { agentSessionAutomaticCompaction } from "./agent-session-compaction.js";
-import type { AgentSessionEvent } from "./agent-session-types.js";
+import {
+  createAssistant,
+  createAssistantResultStream,
+  createResourceLoader,
+  createTestSession,
+  streamMocks,
+  testModel,
+} from "./agent-session-loop.test-support.js";
 import { AgentSession } from "./agent-session.js";
-import { AuthStorage } from "./auth-storage.js";
-import { createExtensionRuntime } from "./extensions/loader.js";
-import type { LoadExtensionsResult, ToolDefinition } from "./extensions/types.js";
-import { ModelRegistry } from "./model-registry.js";
-import type { ResourceLoader } from "./resource-loader.js";
-import { createAgentSession, createAgentSessionForEmbeddedRunner } from "./sdk.js";
+import type { ToolDefinition } from "./extensions/types.js";
 import { SessionManager } from "./session-manager.js";
 import { SettingsManager } from "./settings-manager.js";
-import { createSyntheticSourceInfo } from "./source-info.js";
-
-const testModel: Model = {
-  id: "test-model",
-  name: "Test Model",
-  api: "openai-responses",
-  provider: "test-provider",
-  baseUrl: "https://example.test",
-  reasoning: false,
-  input: ["text"],
-  cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0 },
-  contextWindow: 100,
-  maxTokens: 100,
-};
-
-const sessions: AgentSession[] = [];
-
-function createUsage(contextTokens: number) {
-  return {
-    input: contextTokens,
-    output: 1,
-    cacheRead: 0,
-    cacheWrite: 0,
-    totalTokens: contextTokens + 1,
-    contextUsage: {
-      state: "available" as const,
-      promptTokens: contextTokens,
-      totalTokens: contextTokens + 1,
-    },
-    cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0, total: 0 },
-  };
-}
-
-function createAssistant(
-  activeModel: Model,
-  content: AssistantMessage["content"],
-  stopReason: AssistantMessage["stopReason"] = "stop",
-  contextTokens = 1,
-): AssistantMessage {
-  return {
-    role: "assistant",
-    content,
-    api: activeModel.api,
-    provider: activeModel.provider,
-    model: activeModel.id,
-    usage: createUsage(contextTokens),
-    stopReason,
-    timestamp: Date.now(),
-  };
-}
-
-function createAssistantResultStream(message: AssistantMessage) {
-  const stream = createAssistantMessageEventStream();
-  queueMicrotask(() => {
-    if (message.stopReason === "error" || message.stopReason === "aborted") {
-      stream.push({ type: "error", reason: message.stopReason, error: message });
-    } else {
-      stream.push({ type: "done", reason: message.stopReason, message });
-    }
-    stream.end();
-  });
-  return stream;
-}
-
-const createOverflowAssistant = (activeModel: Model) => ({
-  ...createAssistant(activeModel, [{ type: "text", text: "truncated answer" }], "length", 100),
-  usage: { ...createUsage(100), output: 0 },
-});
-
-const createAutoCompactionSettings = () =>
-  SettingsManager.inMemory({
-    compaction: { enabled: true, reserveTokens: 0, keepRecentTokens: 1 },
-    retry: { enabled: false },
-  });
-
-function mockInvalidThenTextSummary(recoveredText: string) {
-  let requests = 0;
-  streamMocks.streamSimple.mockImplementation((activeModel: Model) => {
-    return createAssistantResultStream(
-      createAssistant(
-        activeModel,
-        ++requests === 1
-          ? [{ type: "thinking", thinking: "internal summary reasoning" }]
-          : [{ type: "text", text: recoveredText }],
-      ),
-    );
-  });
-  return () => requests;
-}
-
-function createResourceLoader(
-  handlers: Map<string, Array<(...args: unknown[]) => Promise<unknown>>> = new Map(),
-): ResourceLoader {
-  const extensionsResult: LoadExtensionsResult = {
-    extensions:
-      handlers.size > 0
-        ? [
-            {
-              path: "<test-extension>",
-              resolvedPath: "<test-extension>",
-              sourceInfo: createSyntheticSourceInfo("<test-extension>", {
-                source: "temporary",
-              }),
-              handlers,
-              tools: new Map(),
-              messageRenderers: new Map(),
-              commands: new Map(),
-              flags: new Map(),
-              shortcuts: new Map(),
-            },
-          ]
-        : [],
-    errors: [],
-    runtime: createExtensionRuntime(),
-  };
-  return {
-    getExtensions: () => extensionsResult,
-    getSkills: () => ({ skills: [], diagnostics: [] }),
-    getPrompts: () => ({ prompts: [], diagnostics: [] }),
-    getThemes: () => ({ themes: [], diagnostics: [] }),
-    getAgentsFiles: () => ({ agentsFiles: [] }),
-    getSystemPrompt: () => undefined,
-    getAppendSystemPrompt: () => [],
-    extendResources: () => {},
-    reload: async () => {},
-  };
-}
-
-function createCompactionHandlers() {
-  return new Map<string, Array<(...args: unknown[]) => Promise<unknown>>>([
-    [
-      "session_before_compact",
-      [
-        async (event: unknown) => {
-          const preparation = (
-            event as {
-              preparation: { firstKeptEntryId: string; tokensBefore: number };
-            }
-          ).preparation;
-          return {
-            compaction: {
-              summary: "condensed history",
-              firstKeptEntryId: preparation.firstKeptEntryId,
-              tokensBefore: preparation.tokensBefore,
-            },
-          };
-        },
-      ],
-    ],
-  ]);
-}
-
-async function createTestSession(
-  options: {
-    model?: Model;
-    settingsManager?: SettingsManager;
-    sessionManager?: SessionManager;
-    resourceLoader?: ResourceLoader;
-    customTools?: ToolDefinition[];
-    contextOverflowRecoveryOwner?: "session" | "caller";
-  } = {},
-) {
-  const model = options.model ?? testModel;
-  const authStorage = AuthStorage.inMemory();
-  authStorage.setRuntimeApiKey(model.provider, "test-api-key");
-  const settingsManager =
-    options.settingsManager ??
-    SettingsManager.inMemory({
-      compaction: { enabled: false },
-      retry: { enabled: false },
-    });
-  const sessionManager = options.sessionManager ?? SessionManager.inMemory();
-  const modelRegistry = ModelRegistry.inMemory(authStorage);
-  modelRegistry.registerProvider(model.provider, {
-    api: model.api,
-    streamSimple: streamMocks.streamSimple,
-  });
-  const sessionOptions = {
-    model,
-    noTools: "builtin" as const,
-    customTools: options.customTools,
-    resourceLoader: options.resourceLoader ?? createResourceLoader(),
-    sessionManager,
-    settingsManager,
-    modelRegistry,
-  };
-  const result = options.contextOverflowRecoveryOwner
-    ? await createAgentSessionForEmbeddedRunner(sessionOptions, {
-        contextOverflowRecoveryOwner: options.contextOverflowRecoveryOwner,
-      })
-    : await createAgentSession(sessionOptions);
-  sessions.push(result.session);
-  return { ...result, settingsManager, sessionManager };
-}
-
-function appendHistory(sessionManager: SessionManager, assistant: AssistantMessage): void {
-  sessionManager.appendMessage({ role: "user", content: "old prompt", timestamp: Date.now() - 2 });
-  sessionManager.appendMessage({ ...assistant, timestamp: Date.now() - 1 });
-}
-
-beforeEach(() => {
-  streamMocks.streamSimple.mockReset();
-});
-
-afterEach(() => {
-  for (const session of sessions.splice(0)) {
-    session.dispose();
-  }
-});
 
 describe("AgentSession loop correctness", () => {
   it("carries the canonical assistant entry id through ordered terminal listeners", async () => {
@@ -304,541 +91,6 @@ describe("AgentSession loop correctness", () => {
     await session.prompt("new prompt");
 
     expect(lifecycleEvents).toEqual(["agent_end", "agent_settled"]);
-  });
-
-  it("manually compacts a completed turn smaller than the retained-token budget", async () => {
-    const sessionManager = SessionManager.inMemory();
-    appendHistory(
-      sessionManager,
-      createAssistant(testModel, [{ type: "text", text: "short answer" }]),
-    );
-    const settingsManager = SettingsManager.inMemory({
-      compaction: { enabled: true, reserveTokens: 0, keepRecentTokens: 10_000 },
-      retry: { enabled: false },
-    });
-    const { session } = await createTestSession({
-      sessionManager,
-      settingsManager,
-      resourceLoader: createResourceLoader(createCompactionHandlers()),
-    });
-
-    const result = await session.compact();
-
-    expect(result.summary).toBe("condensed history");
-    expect(sessionManager.getBranch().at(-1)).toMatchObject({
-      type: "compaction",
-      summary: "condensed history",
-    });
-  });
-
-  it("keeps a successful high-usage response and performs threshold maintenance without retry", async () => {
-    const settingsManager = createAutoCompactionSettings();
-    const compactionEvents: AgentSessionEvent[] = [];
-    streamMocks.streamSimple.mockImplementation((activeModel: Model) =>
-      createAssistantResultStream(
-        createAssistant(activeModel, [{ type: "text", text: "complete answer" }], "stop", 100),
-      ),
-    );
-    const { session } = await createTestSession({
-      settingsManager,
-      resourceLoader: createResourceLoader(createCompactionHandlers()),
-    });
-    session.subscribe((event) => {
-      if (event.type === "compaction_end") {
-        compactionEvents.push(event);
-      }
-    });
-
-    await session.prompt("new prompt");
-
-    expect(streamMocks.streamSimple).toHaveBeenCalledOnce();
-    expect(session.messages).toContainEqual(
-      expect.objectContaining({
-        role: "assistant",
-        content: [{ type: "text", text: "complete answer" }],
-      }),
-    );
-    expect(compactionEvents).toContainEqual(
-      expect.objectContaining({ type: "compaction_end", reason: "threshold", willRetry: false }),
-    );
-  });
-
-  it("does not pre-prompt compact from usage before a zero unavailable marker", async () => {
-    const model = { ...testModel, contextWindow: 1_000 };
-    const sessionManager = SessionManager.inMemory();
-    appendHistory(
-      sessionManager,
-      createAssistant(model, [{ type: "text", text: "old cumulative turn" }], "stop", 950),
-    );
-    sessionManager.appendMessage({ role: "user", content: "CLI prompt", timestamp: Date.now() });
-    sessionManager.appendMessage({
-      ...createAssistant(model, [{ type: "text", text: "usage unavailable" }]),
-      usage: {
-        input: 0,
-        output: 0,
-        cacheRead: 0,
-        cacheWrite: 0,
-        totalTokens: 0,
-        contextUsage: { state: "unavailable" },
-        cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0, total: 0 },
-      },
-    });
-    const settingsManager = SettingsManager.inMemory({
-      compaction: { enabled: true, reserveTokens: 100, keepRecentTokens: 20 },
-      retry: { enabled: false },
-    });
-    const compactionEvents: AgentSessionEvent[] = [];
-    streamMocks.streamSimple.mockImplementation((activeModel: Model) =>
-      createAssistantResultStream(
-        createAssistant(activeModel, [{ type: "text", text: "complete answer" }], "stop", 20),
-      ),
-    );
-    const { session } = await createTestSession({
-      model,
-      sessionManager,
-      settingsManager,
-      resourceLoader: createResourceLoader(createCompactionHandlers()),
-    });
-    session.subscribe((event) => {
-      if (event.type === "compaction_start" || event.type === "compaction_end") {
-        compactionEvents.push(event);
-      }
-    });
-
-    expect(session.messages.at(-1)).toMatchObject({
-      role: "assistant",
-      usage: { contextUsage: { state: "unavailable" } },
-    });
-    expect(session.getContextUsage()?.tokens).toBeLessThan(900);
-    await session.prompt("continue after CLI turn");
-
-    expect(streamMocks.streamSimple).toHaveBeenCalledOnce();
-    expect(compactionEvents).toEqual([]);
-    expect(session.getLastAssistantText()).toBe("complete answer");
-  });
-
-  it("skips threshold maintenance when embedded auto-compaction is disabled", async () => {
-    const settingsManager = SettingsManager.inMemory({
-      compaction: { enabled: false, reserveTokens: 0, keepRecentTokens: 1 },
-      retry: { enabled: false },
-    });
-    const compactionEvents: AgentSessionEvent[] = [];
-    streamMocks.streamSimple.mockImplementation((activeModel: Model) =>
-      createAssistantResultStream(
-        createAssistant(activeModel, [{ type: "text", text: "complete answer" }], "stop", 100),
-      ),
-    );
-    const { session } = await createTestSession({
-      settingsManager,
-      resourceLoader: createResourceLoader(createCompactionHandlers()),
-    });
-    session.subscribe((event) => {
-      if (event.type === "compaction_end") {
-        compactionEvents.push(event);
-      }
-    });
-
-    await session.prompt("new prompt");
-
-    expect(streamMocks.streamSimple).toHaveBeenCalledOnce();
-    expect(compactionEvents).toEqual([]);
-  });
-
-  it("does not retry a high-usage turn terminated by a tool result", async () => {
-    const terminalTool: ToolDefinition = {
-      name: "finish",
-      label: "Finish",
-      description: "finishes the current run",
-      parameters: Type.Object({}),
-      execute: async () => ({
-        content: [{ type: "text", text: "finished" }],
-        details: {},
-        terminate: true,
-      }),
-    };
-    const settingsManager = createAutoCompactionSettings();
-    const compactionEvents: AgentSessionEvent[] = [];
-    streamMocks.streamSimple.mockImplementation((activeModel: Model) =>
-      createAssistantResultStream(
-        createAssistant(
-          activeModel,
-          [{ type: "toolCall", id: "call-finish", name: "finish", arguments: {} }],
-          "toolUse",
-          100,
-        ),
-      ),
-    );
-    const { session } = await createTestSession({
-      settingsManager,
-      resourceLoader: createResourceLoader(createCompactionHandlers()),
-      customTools: [terminalTool],
-    });
-    session.subscribe((event) => {
-      if (event.type === "compaction_end") {
-        compactionEvents.push(event);
-      }
-    });
-
-    await session.prompt("finish now");
-
-    expect(streamMocks.streamSimple).toHaveBeenCalledOnce();
-    expect(compactionEvents).toContainEqual(
-      expect.objectContaining({ type: "compaction_end", reason: "threshold", willRetry: false }),
-    );
-  });
-
-  it("compacts and retries a high-usage length-truncated response", async () => {
-    const settingsManager = createAutoCompactionSettings();
-    const compactionEvents: AgentSessionEvent[] = [];
-    let requestCount = 0;
-    streamMocks.streamSimple.mockImplementation((activeModel: Model) => {
-      requestCount += 1;
-      return createAssistantResultStream(
-        requestCount === 1
-          ? createOverflowAssistant(activeModel)
-          : createAssistant(activeModel, [{ type: "text", text: "complete retry" }]),
-      );
-    });
-    const { session } = await createTestSession({
-      settingsManager,
-      resourceLoader: createResourceLoader(createCompactionHandlers()),
-    });
-    session.subscribe((event) => {
-      if (event.type === "compaction_end") {
-        compactionEvents.push(event);
-      }
-    });
-
-    await session.prompt("long request");
-
-    expect(streamMocks.streamSimple).toHaveBeenCalledTimes(2);
-    expect(compactionEvents).toContainEqual(
-      expect.objectContaining({ type: "compaction_end", reason: "overflow", willRetry: true }),
-    );
-    expect(session.getLastAssistantText()).toBe("complete retry");
-  });
-
-  it("retries a reasoning-only summary once during default auto-compaction", async () => {
-    const settingsManager = createAutoCompactionSettings();
-    const compactionEvents: AgentSessionEvent[] = [];
-    let agentRequests = 0;
-    let summaryRequests = 0;
-    streamMocks.streamSimple.mockImplementation((activeModel: Model, context: Context) => {
-      const isSummary = context.systemPrompt?.includes("context summarization assistant") === true;
-      if (isSummary) {
-        summaryRequests += 1;
-        return createAssistantResultStream(
-          createAssistant(
-            activeModel,
-            summaryRequests === 1
-              ? [{ type: "thinking", thinking: "internal summary reasoning" }]
-              : [{ type: "text", text: "recovered default summary" }],
-          ),
-        );
-      }
-      agentRequests += 1;
-      return createAssistantResultStream(
-        agentRequests === 1
-          ? createOverflowAssistant(activeModel)
-          : createAssistant(activeModel, [{ type: "text", text: "complete retry" }]),
-      );
-    });
-    const { session, sessionManager } = await createTestSession({
-      settingsManager,
-      resourceLoader: createResourceLoader(),
-    });
-    session.subscribe((event) => {
-      if (event.type === "compaction_end") {
-        compactionEvents.push(event);
-      }
-    });
-
-    await session.prompt("long request");
-
-    expect({ agentRequests, summaryRequests }).toEqual({ agentRequests: 2, summaryRequests: 2 });
-    expect(compactionEvents).toContainEqual(
-      expect.objectContaining({ type: "compaction_end", reason: "overflow", willRetry: true }),
-    );
-    const compactionEntry = sessionManager.getBranch().find((entry) => entry.type === "compaction");
-    expect(compactionEntry).toMatchObject({ type: "compaction", fromHook: false });
-    expect(compactionEntry?.summary).toContain("recovered default summary");
-    expect(session.getLastAssistantText()).toBe("complete retry");
-  });
-
-  it("shares invalid-summary recovery with caller-owned automatic compaction", async () => {
-    const sessionManager = SessionManager.inMemory();
-    appendHistory(
-      sessionManager,
-      createAssistant(testModel, [{ type: "text", text: "historical answer to summarize" }]),
-    );
-    const settingsManager = createAutoCompactionSettings();
-    const getSummaryRequests = mockInvalidThenTextSummary("recovered caller-owned summary");
-    const { session } = await createTestSession({
-      sessionManager,
-      settingsManager,
-      resourceLoader: createResourceLoader(),
-    });
-
-    const result = await session[agentSessionAutomaticCompaction]();
-
-    expect(getSummaryRequests()).toBe(2);
-    expect(result.summary).toContain("recovered caller-owned summary");
-    const compactions = sessionManager.getBranch().filter((entry) => entry.type === "compaction");
-    expect(compactions).toHaveLength(1);
-  });
-
-  it("keeps public manual compaction one-shot for invalid summary output", async () => {
-    const sessionManager = SessionManager.inMemory();
-    appendHistory(
-      sessionManager,
-      createAssistant(testModel, [{ type: "text", text: "historical answer to summarize" }]),
-    );
-    const settingsManager = createAutoCompactionSettings();
-    const getSummaryRequests = mockInvalidThenTextSummary("must not be requested");
-    const { session } = await createTestSession({
-      sessionManager,
-      settingsManager,
-      resourceLoader: createResourceLoader(),
-    });
-
-    await expect(session.compact()).rejects.toThrow(
-      "Turn prefix summarization failed: model returned no summary text",
-    );
-
-    expect(getSummaryRequests()).toBe(1);
-    expect(sessionManager.getBranch().some((entry) => entry.type === "compaction")).toBe(false);
-  });
-
-  it("stops default auto-compaction after two invalid summaries", async () => {
-    const settingsManager = createAutoCompactionSettings();
-    const compactionEvents: AgentSessionEvent[] = [];
-    let agentRequests = 0;
-    let summaryRequests = 0;
-    streamMocks.streamSimple.mockImplementation((activeModel: Model, context: Context) => {
-      if (context.systemPrompt?.includes("context summarization assistant")) {
-        summaryRequests += 1;
-        return createAssistantResultStream(
-          createAssistant(activeModel, [
-            { type: "thinking", thinking: `internal summary reasoning ${summaryRequests}` },
-          ]),
-        );
-      }
-      agentRequests += 1;
-      return createAssistantResultStream(createOverflowAssistant(activeModel));
-    });
-    const { session, sessionManager } = await createTestSession({
-      settingsManager,
-      resourceLoader: createResourceLoader(),
-    });
-    session.subscribe((event) => {
-      if (event.type === "compaction_end") {
-        compactionEvents.push(event);
-      }
-    });
-
-    await session.prompt("long request");
-
-    expect({ agentRequests, summaryRequests }).toEqual({ agentRequests: 1, summaryRequests: 2 });
-    expect(compactionEvents).toContainEqual(
-      expect.objectContaining({
-        type: "compaction_end",
-        reason: "overflow",
-        willRetry: false,
-        errorMessage:
-          "Context overflow recovery failed: Turn prefix summarization failed: model returned no summary text",
-      }),
-    );
-    expect(sessionManager.getBranch().some((entry) => entry.type === "compaction")).toBe(false);
-  });
-
-  it.each([1, 2])(
-    "preserves cancellation when aborting during summary attempt %i",
-    async (abortAttempt) => {
-      const settingsManager = createAutoCompactionSettings();
-      const compactionEvents: Array<Extract<AgentSessionEvent, { type: "compaction_end" }>> = [];
-      let agentRequests = 0;
-      let summaryRequests = 0;
-      const created = await createTestSession({
-        settingsManager,
-        resourceLoader: createResourceLoader(),
-      });
-      const { session } = created;
-      streamMocks.streamSimple.mockImplementation((activeModel: Model, context: Context) => {
-        if (context.systemPrompt?.includes("context summarization assistant")) {
-          const summaryAttempt = ++summaryRequests;
-          const stream = createAssistantMessageEventStream();
-          queueMicrotask(() => {
-            if (summaryAttempt === abortAttempt) {
-              session?.abortCompaction();
-            }
-            stream.push({
-              type: "done",
-              reason: "stop",
-              message: createAssistant(activeModel, [
-                { type: "thinking", thinking: `internal summary reasoning ${summaryAttempt}` },
-              ]),
-            });
-            stream.end();
-          });
-          return stream;
-        }
-        agentRequests += 1;
-        return createAssistantResultStream(createOverflowAssistant(activeModel));
-      });
-      session.subscribe((event) => {
-        if (event.type === "compaction_end") {
-          compactionEvents.push(event);
-        }
-      });
-
-      await session.prompt("long request");
-
-      expect({ agentRequests, summaryRequests }).toEqual({
-        agentRequests: 1,
-        summaryRequests: abortAttempt,
-      });
-      expect(compactionEvents).toHaveLength(1);
-      expect(compactionEvents[0]).toMatchObject({
-        type: "compaction_end",
-        reason: "overflow",
-        aborted: true,
-        willRetry: false,
-      });
-      expect(compactionEvents[0]?.errorMessage).toBeUndefined();
-      expect(created.sessionManager.getBranch().some((entry) => entry.type === "compaction")).toBe(
-        false,
-      );
-    },
-  );
-
-  it("does not retry provider errors during default auto-compaction", async () => {
-    const settingsManager = createAutoCompactionSettings();
-    const compactionEvents: AgentSessionEvent[] = [];
-    let agentRequests = 0;
-    let summaryRequests = 0;
-    streamMocks.streamSimple.mockImplementation((activeModel: Model, context: Context) => {
-      if (context.systemPrompt?.includes("context summarization assistant")) {
-        summaryRequests += 1;
-        return createAssistantResultStream({
-          ...createAssistant(activeModel, [], "error"),
-          errorMessage: "provider unavailable",
-        });
-      }
-      agentRequests += 1;
-      return createAssistantResultStream(createOverflowAssistant(activeModel));
-    });
-    const { session, sessionManager } = await createTestSession({
-      settingsManager,
-      resourceLoader: createResourceLoader(),
-    });
-    session.subscribe((event) => {
-      if (event.type === "compaction_end") {
-        compactionEvents.push(event);
-      }
-    });
-
-    await session.prompt("long request");
-
-    expect({ agentRequests, summaryRequests }).toEqual({ agentRequests: 1, summaryRequests: 1 });
-    expect(compactionEvents).toContainEqual(
-      expect.objectContaining({
-        type: "compaction_end",
-        reason: "overflow",
-        willRetry: false,
-        errorMessage:
-          "Context overflow recovery failed: Turn prefix summarization failed: provider unavailable",
-      }),
-    );
-    expect(sessionManager.getBranch().some((entry) => entry.type === "compaction")).toBe(false);
-  });
-
-  it("leaves reactive overflow recovery to the caller when configured", async () => {
-    const settingsManager = createAutoCompactionSettings();
-    const compactionEvents: AgentSessionEvent[] = [];
-    streamMocks.streamSimple.mockImplementation((activeModel: Model) =>
-      createAssistantResultStream({
-        ...createAssistant(activeModel, [], "error", 100),
-        errorMessage: "400 Your input exceeds the context window of this model",
-      }),
-    );
-    const { session } = await createTestSession({
-      settingsManager,
-      resourceLoader: createResourceLoader(createCompactionHandlers()),
-      contextOverflowRecoveryOwner: "caller",
-    });
-    session.subscribe((event) => {
-      if (event.type === "compaction_end") {
-        compactionEvents.push(event);
-      }
-    });
-
-    await session.prompt("long request");
-
-    expect(streamMocks.streamSimple).toHaveBeenCalledOnce();
-    expect(compactionEvents).toEqual([]);
-    expect(session.messages.at(-1)).toMatchObject({
-      role: "assistant",
-      stopReason: "error",
-      errorMessage: "400 Your input exceeds the context window of this model",
-    });
-  });
-
-  it("keeps threshold maintenance session-owned when the caller owns overflow recovery", async () => {
-    const settingsManager = createAutoCompactionSettings();
-    const compactionEvents: AgentSessionEvent[] = [];
-    streamMocks.streamSimple.mockImplementation((activeModel: Model) =>
-      createAssistantResultStream(
-        createAssistant(activeModel, [{ type: "text", text: "complete answer" }], "stop", 100),
-      ),
-    );
-    const { session } = await createTestSession({
-      settingsManager,
-      resourceLoader: createResourceLoader(createCompactionHandlers()),
-      contextOverflowRecoveryOwner: "caller",
-    });
-    session.subscribe((event) => {
-      if (event.type === "compaction_end") {
-        compactionEvents.push(event);
-      }
-    });
-
-    await session.prompt("long request");
-
-    expect(streamMocks.streamSimple).toHaveBeenCalledOnce();
-    expect(compactionEvents).toContainEqual(
-      expect.objectContaining({ type: "compaction_end", reason: "threshold", willRetry: false }),
-    );
-  });
-
-  it("delivers a pending prompt immediately after pre-prompt compaction", async () => {
-    const sessionManager = SessionManager.inMemory();
-    appendHistory(
-      sessionManager,
-      createAssistant(testModel, [{ type: "text", text: "old answer" }], "stop", 100),
-    );
-    const settingsManager = SettingsManager.inMemory({
-      compaction: { enabled: true, reserveTokens: 0, keepRecentTokens: 1 },
-      retry: { enabled: false },
-    });
-    const requests: Context[] = [];
-    streamMocks.streamSimple.mockImplementation((activeModel: Model, context: Context) => {
-      requests.push(context);
-      return createAssistantResultStream(
-        createAssistant(activeModel, [{ type: "text", text: "new answer" }]),
-      );
-    });
-    const { session } = await createTestSession({
-      sessionManager,
-      settingsManager,
-      resourceLoader: createResourceLoader(createCompactionHandlers()),
-    });
-    const continueRun = vi.spyOn(session.agent, "continue");
-
-    await session.prompt("pending prompt");
-
-    expect(continueRun).not.toHaveBeenCalled();
-    expect(requests).toHaveLength(1);
-    expect(JSON.stringify(requests[0]?.messages)).toContain("pending prompt");
   });
 
   it("drains a follow-up queued by an agent-end handler", async () => {
