+++ src/auto-reply/reply/agent-runner.misc.runreplyagent.test.ts
@@ -557,6 +557,8 @@ describe("runReplyAgent auto-compaction token update", () => {
       sessionId: "session",
       updatedAt: Date.now(),
       totalTokens: 200_000,
+      totalTokensFresh: true,
+      totalTokensVersion: 1 as const,
     };
     await seedSessionStore({ storePath, sessionKey, entry: sessionEntry });
     compactState.compactEmbeddedAgentSessionMock.mockRejectedValueOnce(new GatewayDrainingError());
@@ -2779,6 +2781,8 @@ describe("runReplyAgent fallback reasoning tags", () => {
         sessionId: "session",
         updatedAt: Date.now(),
         totalTokens: 1_000_000,
+        totalTokensFresh: true,
+        totalTokensVersion: 1,
         compactionCount: 0,
       },
     });
