+++ src/auto-reply/reply/commands-status.test.ts
@@ -2027,6 +2027,8 @@ describe("buildStatusReply subagent summary", () => {
         sessionId: "sess-status-codex-context",
         updatedAt: 0,
         totalTokens: 25_000,
+        totalTokensFresh: true,
+        totalTokensVersion: 1,
       },
       sessionKey: "agent:main:main",
       parentSessionKey: "agent:main:main",
@@ -2073,6 +2075,8 @@ describe("buildStatusReply subagent summary", () => {
         sessionId: "sess-status-codex-stale-context",
         updatedAt: 0,
         totalTokens: 181_000,
+        totalTokensFresh: true,
+        totalTokensVersion: 1,
         contextTokens: 400_000,
       },
       sessionKey: "agent:main:main",
