+++ src/auto-reply/reply/reply-state.test.ts
@@ -284,7 +284,7 @@ describe("shouldRunMemoryFlush", () => {
   it("skips when under threshold", () => {
     expect(
       shouldRunMemoryFlush({
-        entry: { totalTokens: 10_000 },
+        entry: { totalTokens: 10_000, totalTokensFresh: true, totalTokensVersion: 1 },
         contextWindowTokens: 100_000,
         reserveTokensFloor: 20_000,
         softThresholdTokens: 10_000,
@@ -295,7 +295,7 @@ describe("shouldRunMemoryFlush", () => {
   it("triggers at the threshold boundary", () => {
     expect(
       shouldRunMemoryFlush({
-        entry: { totalTokens: 85 },
+        entry: { totalTokens: 85, totalTokensFresh: true, totalTokensVersion: 1 },
         contextWindowTokens: 100,
         reserveTokensFloor: 10,
         softThresholdTokens: 5,
@@ -308,6 +308,8 @@ describe("shouldRunMemoryFlush", () => {
       shouldRunMemoryFlush({
         entry: {
           totalTokens: 90_000,
+          totalTokensFresh: true,
+          totalTokensVersion: 1,
           compactionCount: 2,
           memoryFlush: { kind: "succeeded", compactionCount: 2 },
         },
@@ -321,7 +323,12 @@ describe("shouldRunMemoryFlush", () => {
   it("runs when above threshold and not flushed", () => {
     expect(
       shouldRunMemoryFlush({
-        entry: { totalTokens: 96_000, compactionCount: 1 },
+        entry: {
+          totalTokens: 96_000,
+          totalTokensFresh: true,
+          totalTokensVersion: 1,
+          compactionCount: 1,
+        },
         contextWindowTokens: 100_000,
         reserveTokensFloor: 5_000,
         softThresholdTokens: 2_000,
@@ -337,14 +344,23 @@ describe("shouldRunMemoryFlush", () => {
     };
 
     for (const entry of [
-      { totalTokens: 95_000, compactionCount: 1 },
       {
         totalTokens: 95_000,
+        totalTokensFresh: true as const,
+        totalTokensVersion: 1 as const,
+        compactionCount: 1,
+      },
+      {
+        totalTokens: 95_000,
+        totalTokensFresh: true as const,
+        totalTokensVersion: 1 as const,
         compactionCount: 2,
         memoryFlush: { kind: "succeeded" as const, compactionCount: 1 },
       },
       {
         totalTokens: 95_000,
+        totalTokensFresh: true as const,
+        totalTokensVersion: 1 as const,
         compactionCount: 3,
         memoryFlush: { kind: "succeeded" as const, compactionCount: 2 },
       },
